# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Polygon DID-Method Demo",
        "description": "A NodeJs application to polygon demo",
        "license": {
            "name": "MIT"
        }
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/update-traceability-did/{did}/{privateKey}": {
            "post": {
                "tags": [
                    "Update traceability DID"
                ],
                "summary": "traceability test function",
                "description": "traceability test function",
                "parameters": [{
                    "name": "did",
                    "in": "path",
                    "required": true
                }, {
                    "name": "privateKey",
                    "in": "path",
                    "required": true
                }],
                "produces": [
                    "string"
                ],
                "responses": {
                    "200": [{
                        "description": "OK",
                        "content":[{
                            "image/png":[{
                                "type": "string",
                                "format": "binary"
                            }]
                         }]
                    }]
                }
            }
        },
...
    }
}